refactor(plugin): Reformat all components to follow new standard
This commit introduces a breaking change that changes the names of all components and makes it impossible to update Grasshopper definitions with old LadybugPlus components in them.

A new component to sync Grasshopper files with the version in the toolbar has been added along with a new exporter component.

A few other fixes have been included in this update like the license string at the top as well as the separation of all dependency imports into different try/except loops.