feat(legendpar): Add component for LegendParameterCategorized
This commit adds a new component to make legend parameters that sort the data into categories instead if using the typical min, max, and segment_count.

It also adds a little more structure to the automatically-downloaded weather files so that we don't inadvertently make a mess of the new directory.