feat(data): Add components for DataCollection arithmetic
This commit adds components for doing arithmetic with data collections.

It also adds a component for constructing a custom DataType, which should address an issue that @mostapharoudsari raised earlier when he was trying to visualize dynamic states on the hourly plot.

Lastly, I realized that I sent the "orient to camera" component to the wrong place last time so i fixed this.