feat(versioner): Add versioner component for updating installed version
I also made some tweaks to the wind rose component, one of which was changing the name from "Wind Rose Plot" to just "Wind Rose". If the word "Plot" was helping us distinguish the component from another, then I would have kept but but, otherwise, brevity is the soul of wit.