feat(geometry): Add component for view from sun
This commit adds a component to set the view from a sun vector.  It also strips out all of the Rhino-specific dependencies from the "Orient to Camera" component.

Lastly, I put in some checks to ensure that all of the climate graphics get produced at the same scale as Rhino units change.