feat(geometry): Add direct Sun Study Component
This commit adds a component to run direct sun studies with Rhino geometry.  It makes use of the highly reusable functions within ladybug-core and ladybug-rhino to do so such that the whole process happens in less than 50 lines of Python.  I expect the radiation and view components to be similar as they will make use of the same functions.