feat(geometry): Add component to perform view analysis
This commit adds a component to analyze the "View Percent" to the outdoors and the sky.  This will be distinguished from a future "Visibility Percent" component, which will quantify the fraction of an object that is visible.

I'm also changing the name of the shadow study component  to be "Direct Sun Hours".