fix(visualize): Fix bugs in sky dome
There were a fee bugs with the dome like the fact that it was mutating the input legend parameters and the horizontal legends did not look correct.

Also, I couldn't think of a use case where I really wanted to have the patch solid angles but I could think of a ton where I want the mesh face values so I swapped the patch_angles output for mesh_values.  If people ever need the solid angles for whatever reason, they are easily accessible on the API  layer.