fix(visualize): Ensure zero-speed hours are filtered in wind rose
This commit puts back some important code that was recently deleted, which ensures zero-speed hours are filtered from all wind roses when speed is connected.

I also removed a lot of unused variables. Lastly, the lines of the plot were really busy and so I made the `windrose_line` hidden by default. I found this makes the data a lot more readable in terms of the frequency lines and people can always preview the `windrose_line` by hooking it up to a separate component if they want.