fix(geometry): Correct the calculation to align with the units
It seems there was no good reason to multiply the results by the area of the study region. It was just making the results harder to understand. Now, we can understand the total of the results in terms of the net degree-days improved. That is, the cooling degree-days helped minus the heating degree-days harmed. And the total result falls within the limits of the maximum possible levels of helping/harming.